backend/requirements.txt
# FastAPI 框架
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18

# 数据验证
pydantic>=2.10.0
pydantic-settings>=2.7.0

# 文件处理
python-docx==1.2.0
PyMuPDF==1.27.2.2
Pillow>=11.0.0
pdf2image==1.17.0
pywin32==311; sys_platform == 'win32'  # Windows: 用于处理 .doc 文件

# HTTP 客户端
httpx>=0.28.0

# GPU 显存（侧栏 /health/services；NVML，不依赖 PATH 中的 nvidia-smi）
nvidia-ml-py>=12.535.133
aiofiles>=24.1.0

# HaS Image（YOLO11 微服务 has_image_server.py，端口 8081）
ultralytics>=8.3.0

# OCR（PaddleOCR-VL 微服务 ocr_server.py）
# 推荐：先安装 GPU 版 Paddle，再装本文件，避免默认装上 CPU 版 paddlepaddle。
# Windows + NVIDIA：在项目根目录执行 scripts\install_paddle_gpu.ps1，然后:
#   pip install -r requirements.txt
# 验证: python -c "import paddle; print(paddle.get_device())"  # 应为 gpu:0
paddleocr[doc-parser]>=2.7.0

# 工具库
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4

# 可观测性
python-json-logger>=2.0.0
prometheus-client>=0.20.0

# 安全（可选：文件加密需要 cryptography，病毒扫描需要 pyclamd + ClamAV daemon）
cryptography>=42.0.0
# pyclamd>=0.4.0  # 按需安装，需配合 ClamAV daemon

# 开发工具
python-dotenv>=1.0.1
pytest>=8.0.0
